Send pleaseObserveBefore{Play,Discard} messages to the active player too.
The new method server.activeCard() allows even the currently active
player to see the identity of the card he's playing or discarding.
This makes it meaningful to give the pleaseObserveBeforePlay()
notification to the active bot as well as the others, and allows
the bot writer to consolidate some similar code into that notification
method. Notice how SimpleBot got even simpler after this change.