Make SimpleBot a little smarter about deduction.
Because SimpleBot gives hints only about playable cards,
you can actually infer the value of a card from a color hint.
For example, if SimpleBot says "this card is red", and there
is currently a 2 on top of the red pile, you know that the
named card must be the red 3.

This extra deduction power also requires that we use the
right logic inside wipeOutPlayables(), so that Alice's
play of the orange 3 doesn't wipe out Bob's knowledge that
his red 3 is playable (despite their both being 3s).

After these two complementary patches, SimpleBot's win
percentage is basically unchanged.