ValueBot now avoids giving hints misconstruable as warnings.
Sometimes Alice wants to give a helpful hint to Bob; for
example, if he has three 2s that are all playable right now.
But if Bob has no currently known-playable cards, and one
of those 2s is in Bob's "oldest non-valuable card" slot,
then Bob will misconstrue Alice's hint as a warning not to
discard that 2.

The fix is to get Alice not to give hints that could be
misconstrued in that way, and simultaneously to tighten up
the conditions under which Bob will interpret the hint as
a warning. (For example, if there are no 2s in the discard
pile, then Bob knows his 2 can't be valuable; therefore,
Alice must be telling him that it's playable.)

This patch bumps ValueBot's average score from 18.08 to 18.11.