Allow ValueBot to give hints relying more on deduction.
Up to this point, ValueBot never gave a "helpful" hint that involved
naming an unplayable card. However, it's pretty common that the hintee
is already aware of the values of some of his cards. For example, if
Alice tells Bob that his oldest card is red, and Bob already knows
that it's a 5, then Bob can deduce for himself whether the red 5 is
playable; Alice doesn't need to avoid giving a clue about red cards
just because the red 5 might not be playable right now.

This patch bumps ValueBot's average score from 18.21 to 18.58,
a boost of more than 0.35 points per game!