Improve CheatBot (avg score 24.50).
CheatBot now tries to enable followup plays from its partners;
for example, if Alice has the blue 1 and the red 2, and Bob has
the red 3, Alice will prefer to play the red 2 even though the
blue 1 is "lower" (assuming that Bob doesn't also have the blue 2).

Also, make the "temporize early just to see what comes up" heuristic
a little more continuous. It's still not a great heuristic, in that
it's hard to explain rationally why this exact heuristic should help;
but it does boost the average score by a few hundredths of a point.
More work is needed.