Rename FinesseBot to HolmesBot. Improve its average from 18.9 to 19.2.
"HolmesBot" is a much better name for a bot that makes use of deductive
reasoning, and that's the main way this bot differs from ValueBot.
It now keeps track not only of the publicly visible played/discarded cards,
but also of the cards that have been definitely identified in other
players' hands. So, for example, if one red 3 is in the discard pile
and the other is in Bob's hand, Alice can conclude that her 3 is
definitely not red (and this may later allow her to deduce that
whatever color her 3 may be, it is definitely playable).

Ultimately I'll want to switch over to an exhaustive search like
Dancing Links, rather than this cascade of ad-hoc deductive rules.
But for now it's pretty good.