Take risks near the end of the game.
This heuristic needs a lot of work, but the basic idea is that
if we're resigned to discarding a card anyway, maybe we should
just play some unknown card and see what happens. Maybe it'll
turn out to be playable, and we'll gain a point!

On the other hand, maybe we want to discard in order to get a
hint-stone back, in which case this is just getting in the way
of our usual strategy.

This patch bumps HolmesBot's average score from 19.23 to 19.26.