Improve HolmesBot's granularity.
Rather than keeping track of "can this card be color K" and "can this
card be value V" in separate arrays, it makes more mathematical sense
to keep track of the single question "can this card be K,V" in a
two-dimensional array. This allows a very slightly finer granularity
of deduction; for example, maybe a certain card could be either the
red 2 or the green 3, and therefore must be playable; whereas merely
knowing that the card was "red or green, and 2 or 3" would not have
allowed for the same deduction.

This boosts HolmesBot's average score from 19.33 to 19.34; hardly
worth mentioning, yet.