Replace CheatBot's "rescue" heuristic with a better rule.
The better rule is: If there are absolutely no playable cards visible,
then you MUST discard a worthless card (if possible) in order to cycle
new cards into play. Temporizing at that point is counter-productive,
as it does nothing but reduce the number of hint-stones.

After implementing this rule, it turned out that CheatBot's old rule
"Attempt to rescue the player to your left if he has nothing to play"
caused more harm than help, so I've gotten rid of it.

This patch improves CheatBot's average score from 24.50 to 24.67.
In about 77.4 percent of those games, CheatBot achieves a perfect 25.

The "temporize early" heuristic still helps quite a bit, unfortunately,
so I've kept it. (Setting shouldTemporizeEarly to FALSE results in an
average score of only 24.29, and only 39.6 percent of those games are
perfect. Always TRUE results in 24.56, 72.1 percent perfect.)