Improve HolmesBot's range of available hints.
HolmesBot avoids giving any hint that could be misconstrued as indicating
that some card was playable when in fact it's not. However, if a certain
card's identity would be perfectly known after the hint, then we don't
need to worry about anyone misconstruing its playability. With this patch,
we also observe that if the card is already known to be worthless, then
we don't need to worry about it either.

This improves HolmesBot's average score from 19.34 to 19.41.

For the first time, HolmesBot achieves a perfect score in a handful
of games! (The percentage is still insignificant: 3 or 4 perfect games
per million played.)