Allow HolmesBot and SmartBot to deduce valuableness.
HolmesBot and SmartBot use deduction to narrow down the possible
identities of each card, so that for example they could know that Alice's
third card is "either a green 2 or a red 4". If green 2s and red 4s are
both worthless, the bots will further deduce the card's isWorthless flag;
ditto if the card must be isPlayable. However, formerly these bots did not
bother to deduce the isValuable flag. This turns out to be a big win!

This patch improves SmartBot's average score from 19.51 to 19.66,
and HolmesBot's average score from 19.41 to 19.56.