Improve SmartBot's observation of value hints, and its use of mulligans.
If a value hint is being given to someone who is not sitting to the
immediate left of the hint-giver, then it cannot be a "valuable card"
warning (at least, under SmartBot's current invariants). SmartBot now
recognizes this fact and incorporates it into its deductions.

It turns out that SmartBot is getting so good that it now really doesn't
want to use mulligans until the last three turns of the game. This
trivial change boosts its average by about 0.01 ppg.