Minor refactors to CheatBot; recognize a new class of worthless cards.
There were a few vestigial organs in CheatBot (for example, the fact
that endgameNoMoreDiscarding was a global variable); these have now been
cleaned up.

maybeDiscardWorthlessCard() and maybeDiscardDuplicateCard() are now
two separate functions, since duplicates are not technically "worthless".

maybeDiscardWorthlessCard() now recognizes that if the red stack is
currently showing "2", and all the red 4s are in the discard pile,
then the red 5 is worthless (as are the red 1s and 2s, for a different
reason; but red 3s are still worth something). This boosts CheatBot's
average insignificantly.

maybeEnablePlay() uses a new heuristic: If the other player already has
a playable card, then we shouldn't bother playing any card that enables
a play by him; he's already taken care of. (There is a corollary: that
if the play he has in mind is the red 2, we shouldn't needlessly play
a red 2 of our own. This corollary is not yet implemented.)