Improve SmartBot's original "smart" feature.
SmartBot is smart because it prefers to play cards that aren't publicly known
to be playable. However, it was just picking the oldest such card in its
hand, rather than picking the lowest such card, in the unlikely case that
its hand contained two different privately-known-playable cards. This is
now fixed, with a couple of TODOs to improve its heuristic in general
(for the more common case that it has two or more playable cards; playing
the lowest one may not always be the best move).

This tweak improves SmartBot's average score by about 0.01 ppg, if that.