Improve SmartBot's sense of when a value hint is a warning.
Before this change, SmartBot would avoid giving its left-hand neighbor a
color hint that included any non-playable cards, unless the non-playable
cards were already publicly known to be worthless or publicly known to
be a specific value. Now SmartBot is allowed to give such hints, if the
non-playable cards can then be deduced not to be playable. For example,
if Bob's third card is "either the green 2 or the red 3", Alice might
safely tell him that it is red, even if the red 3 is not playable, because
she can rely on Bob's deductive reasoning to conclude that of course it's
not playable.