Fix some bugs in SmartBot.
In several places, I'd misused the "playableValue = pile.size()+1" idiom
in a way that mishandled a completed pile. For example, the couldBePlayable()
method indexed off the end of the cantBe_ array, and pleaseObserveColorHint()
constructed a Card with value 6.

Introduce an isWorthless() method to be symmetric with isValuable(), and
rewrite isValuable() in terms of !isWorthless().

These bugfixes improve SmartBot's average score by a little more than 0.01 ppg.