Fix a bug in Hanabi::Server.
The server forgot to clear the discard pile (a vector of Cards)
before each game, which meant that calling server.discards() in
any game after the first would return a vector containing too
many cards. None of my bots relied on this behavior, so I didn't
catch it until now.