SmartBot now prefers plays over hints, during the endgame.
On the "easiest possible" deck, SmartBot was leaving a point on the table
by preferring maybeGiveValuableHint() over maybePlayLowestCard() at one
point near the end of the game. When we're definitely in the endgame, it
doesn't make sense to give a hint (maybe scoring a point) in preference
to playing one's own lowest card (definitely scoring a point).

This boosts SmartBot's average score by 0.07 ppg.
SmartBot's current win percentage is 6.7-% perfect, and I've added that
information to the README for quick reference going forward.