SmartBot is now faster! No functional change.
Finally refactor the whole isPlayable/isValuable/isWorthless
bookkeeping to happen "instantaneously": as soon as SmartBot
deduces that a card is fooable, it goes through and sets the
cantBe_ bits on all unfooable color/value pairs. This means
that the update() method itself no longer cares about the
incoming values of isPlayable and isValuable, which in turn
means that wipeOutPlayables() is obsolete.

This gives a 27.4% speedup when built with FAST=1 OPENMP=0,
and has no functional effect.