Refactoring SmartBot in preparation for trivalues.
Be consistent about the alphabetical ordering of "Playable",
"Valuable", and "Worthless".

Add a method bot.isPlayable(server,card) for symmetry with
isValuable() and isWorthless().

Rename couldBeValuable() to couldBeValuableWithValue(),
for clarity.

Refactor CardKnowledge::update() to enable trivalue logic
instead of boolean logic: in other words, we want to be able
to deduce when a card is definitely NOT p/v/w, just as
much as we want to know when it definitely IS p/v/w.
In this patch, we end up throwing away that information as
soon as we calculate it (i.e., we flatten our trivalue
NO/MAYBE/YES into a boolean NO_OR_MAYBE/YES); but this
enables us to incrementally refactor SmartBot to the point
where the information will not be thrown away.

No functional change. The new algorithm for update() is
significantly slower, but once we finish the refactor, it'll
be worth it.