Implement 5-card hands for 2 and 3 players.
Geez, Hanabi has so many fiddly rules! I guess I was wrong when
I wrote that this project "now implements all the rules of Hanabi"...
but maybe *now* it's accurate. The patch to support hands bigger than
four cards was a lot more invasive than I wanted it to be, but it
seems to be done now. This totally ruins my perfect percentage for
three players: the game seems to be harder to win with bigger hands,
although I don't know how much of that is actually the deck running
out sooner, and how much is all my finely tuned ad-hoc heuristics
being wrecked.

Refactored the README to include stats for all numbers of players.