SmartBot now tries to discard cards others don't know it knows.
If Alice deduces that her Nth card is worthless, but this fact
is not public knowledge (i.e., Bob doesn't know that Alice knows),
then Alice should prefer to discard that Nth card rather than
holding onto it and possibly causing Bob to waste a hint on it.

This has an insignificant effect on SmartBot's win rate.