Implement the "discard finesse" for SmartBot.
As seen in this BoardGameGeek forum thread:
https://boardgamegeek.com/thread/1326902/discard-finesse

If everyone at the table knows that Alice's second card is
the (playable) red 2, but Alice looks around the table and sees
that Bob has the other red 2 in his newest position, then it is
beneficial for Alice to transmit that information to Bob by
discarding her own red 2. Bob, seeing Alice unexpectedly throw
away a playable card, will conclude that he himself must be the
holder of the other red 2 (as his newest card).

This boosts SmartBot's win percentage by a few tenths of a percent.