Try a new strategy in bestHintForPlayer().
This strategy tries to maximize the reduction in the entropy of
the target player's hand. This might actually be counterproductive
when massive strategy is an option, because it reduces the usefulness
of future massive hints! And indeed, we see the average score for
4-player games dipping even as the average score for 2p and 3p games
rises.

This strategy might be a bigger improvement in old SmartBot, without
massive hints. However, this strategy does still enforce the rule that
"The newest of the named cards must be playable (if not already known)."
It might be even more useful to relax that rule. For example, we could
give a helpful hint that names *no* playable cards, as long as it's
not misleading; but then we'd have to assign such hints nonzero fitness,
which means we'd have to stop using `hint.fitness` as a side channel
in `maybeGiveValuableWarning`. We'd have to split out a bit for
`hint.providesPlayableCard` and/or `hint.providesWorthlessCard`
(since either of those would stop the warned player from discarding
a maybe-valuable chop card).

Actually, `maybeGiveValuableWarning` should be rewritten to find its
own hints instead of relying on `bestHintForPlayer`. The hint that
provides a playable/worthless card (or by deduction reveals
`chopCard.valuable() == YES`) might not be the "best" hint by
the usual metric.