Make SmartBot a little smarter about helpful hints.
In `maybeGiveHelpfulHint`, compute fitness based on overall reduction
in the possibility space, not just in the absolute number of playable
cards explicitly identified. Also, if a color or value hint can reveal
a playable card through the *negative* space (i.e. if a card becomes
known-playable as soon as it is seen *not* to be touched), SmartBot
now considers this a good helpful hint.