InfoBot: Improve HintStrategySetPacking.
It should place every color and every value into *some* set
in the partition; otherwise it will never consider those hints
and possibly miss the chance to give some overt information
that could help.

Unfortunately this does not gain back any of the win-ratio
that was lost by moving to HintStrategySetPacking in the
first place.